# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Nöbetçi Eczane API
  version: 0.1.0
  description: |
    Nöbetçi eczane bulma uygulamasının API sözleşmesi.

    **Bu dosya tek doğru kaynaktır.** Frontend tipleri buradan üretilir:
    `npm run gen:types` → `src/shared/api-types.d.ts`

    Elle tip yazılmaz. Önce bu dosya değişir, sonra tipler üretilir, sonra kod yazılır.

    Veri kaynağı ve tazeleme davranışı için: `docs/adr/001-veri-kaynagi.md`,
    `docs/adr/003-veri-tazeleme.md`
servers:
  - url: http://localhost:8787
    description: Yerel geliştirme (wrangler dev)
  - url: https://nobetci-eczane.workers.dev
    description: Production
...
paths:
...
  /api/districts:
    get:
      tags: [public]
      operationId: listDistricts
      summary: İlçe listesi ve ilçe başına nöbetçi sayısı
      description: |
        Konum izni reddedildiğinde gösterilen ilçe seçim ekranını besler.
        Nöbetçi sayısı o an nöbette olan eczaneleri sayar.
      parameters:
        - $ref: "#/components/parameters/CityCode"
      responses:
        "200":
          description: İlçe listesi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DistrictList" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/pharmacies/on-duty:
    get:
      tags: [public]
      operationId: listOnDutyPharmacies
      summary: Nöbetçi eczaneler
      description: |
        İki kullanım biçimi vardır ve **tam olarak biri** seçilmelidir:

        - **Konuma göre:** `lat` + `lng` verilir. Sonuç mesafeye göre artan sıralanır,
          her kayıtta `distanceM` ve `etaMin` bulunur.
        - **İlçeye göre:** `district` verilir. Sonuç eczane adına göre sıralanır,
          `distanceM` ve `etaMin` `null` olur.

        `etaMin` haversine mesafesinden türetilmiş **tahmindir**; gerçek trafik verisi
        içermez. Gerekçe: `docs/adr/002-harita-yol-tarifi.md`

        Nöbeti bitmiş eczaneler varsayılan olarak dönmez; `includeExpired=true` ile
        `status: "closed"` şeklinde döner (tasarımdaki soluk satırlar).

        **İl konumdan çözülür.** `lat`/`lng` verilip `city` verilmezse sunucu en yakın
        eczanenin ilini seçer; hangi ili seçtiği yanıtın `cityCode` alanındadır.
        Kullanıcı ili elle değiştirdiğinde `city` gönderilir ve konumdan çözüm yapılmaz.

        Veri o il için henüz çekilmemişse `items` boş ve `stale` `true` döner —
        **hata değildir**, arayüz boş durum ekranını gösterir.
      parameters:
        - name: lat
          in: query
          description: Kullanıcı enlemi. `lng` ile birlikte verilir, `district` ile birlikte verilemez.
          schema: { type: number, format: double, minimum: 35.8, maximum: 42.2 }
          example: 36.8969
        - name: lng
          in: query
          description: Kullanıcı boylamı. `lat` ile birlikte verilir, `district` ile birlikte verilemez.
          schema: { type: number, format: double, minimum: 25.6, maximum: 44.9 }
          example: 30.7133
        - name: district
          in: query
          description: TÜİK ilçe kodu. `lat`/`lng` ile birlikte verilemez.
          schema: { type: string, pattern: "^[0-9]{4}$" }
          example: "0715"
        - $ref: "#/components/parameters/CityCode"
        - name: limit
          in: query
          description: Döndürülecek azami kayıt sayısı.
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: includeExpired
          in: query
          description: Nöbeti bitmiş eczaneleri de döndür.
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Nöbetçi eczane listesi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PharmacyList" }
              examples:
                konuma-gore:
                  summary: Konuma göre, taze veri
                  value:
                    stale: false
                    dataAsOf: "2026-08-07T05:05:12Z"
                    nextRotationAt: "2026-08-08T05:00:00Z"
                    cityCode: 7
                    items:
                      - id: "antalya-eo:8680001234567"
                        name: "Deniz Eczanesi"
                        phone: "+902422371414"
                        address: "Tahılpazarı Mah. Ali Çetinkaya Cad. No:41/A"
                        districtCode: "0715"
                        districtName: "Muratpaşa"
                        lat: 36.8862
                        lng: 30.7056
                        distanceM: 1240
                        etaMin: 4
                        dutyStart: "2026-08-07T05:00:00Z"
                        dutyEnd: "2026-08-08T05:00:00Z"
                        status: "open"
                        minutesUntilClose: 318
                bayat:
                  summary: Senkron başarısız, dünkü veri servis ediliyor
                  value:
                    stale: true
                    dataAsOf: "2026-08-06T05:05:09Z"
                    nextRotationAt: "2026-08-08T05:00:00Z"
                    cityCode: 7
                    items: []
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    CityCode:
      name: city
      in: query
      description: |
        Plaka kodu (81 ilin tamamı). `lat`/`lng` verildiğinde **gerekmez**: sunucu
        ili konumdan çözer ve hangi ili seçtiğini yanıttaki `cityCode` ile bildirir.
        Konum yoksa varsayılan 7 (Antalya).
      schema: { type: integer, minimum: 1, maximum: 81, default: 7 }
  responses:
    BadRequest:
      description: Geçersiz istek parametresi
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthorized
                - not_found
                - city_not_supported
                - rate_limited
                - source_unavailable
                - dispatch_failed
                - not_configured
                - internal
            message:
              type: string
              description: Kullanıcıya gösterilebilir Türkçe mesaj.
            details:
              type: array
              description: Alan bazlı doğrulama hataları.
              items:
                type: object
                required: [path, message]
                additionalProperties: false
                properties:
                  path: { type: string }
                  message: { type: string }
...
    District:
      type: object
      required: [code, name, onDutyCount]
      additionalProperties: false
      properties:
        code: { type: string, pattern: "^[0-9]{4}$", example: "0715" }
        name: { type: string, example: "Muratpaşa" }
        onDutyCount: { type: integer, minimum: 0, example: 6 }
    DistrictList:
      type: object
      required: [cityCode, stale, dataAsOf, items]
      additionalProperties: false
      properties:
        cityCode: { type: integer, example: 7 }
        stale: { type: boolean }
        dataAsOf: { type: [string, "null"], format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/District" }
    Pharmacy:
      type: object
      required:
        [
          id,
          name,
          phone,
          address,
          districtCode,
          districtName,
          lat,
          lng,
          distanceM,
          etaMin,
          dutyStart,
          dutyEnd,
          status,
          minutesUntilClose,
        ]
      additionalProperties: false
      properties:
        id:
          type: string
          description: "`{sourceKey}:{sourceId}` — kaynaklar arası çakışmayan kalıcı kimlik."
          example: "antalya-eo:8680001234567"
        name: { type: string, example: "Deniz Eczanesi" }
        phone:
          type: [string, "null"]
          description: E.164 formatında. `tel:` bağlantısında doğrudan kullanılır.
          example: "+902422371414"
        address: { type: string }
        districtCode: { type: string, pattern: "^[0-9]{4}$" }
        districtName: { type: string }
        lat:
          type: [number, "null"]
          description: Koordinat kaynaktan ayrı bir istekle geliyor; henüz bilinmiyor olabilir.
        lng: { type: [number, "null"], format: double }
        distanceM:
          type: [integer, "null"]
          description: Kuş uçuşu mesafe (metre). İlçeye göre sorguda `null`.
        etaMin:
          type: [integer, "null"]
          description: |
            **Tahmini** sürüş süresi (dakika). Mesafeden türetilir, trafik içermez.
            Arayüzde `~` ile gösterilir. İlçeye göre sorguda `null`.
        dutyStart: { type: string, format: date-time }
        dutyEnd: { type: string, format: date-time }
        status:
          type: string
          enum: [open, closing_soon, closed]
          description: |
            `closing_soon`: kapanışa 120 dakikadan az kaldı — tasarımda turuncu varyant.
            `closed`: nöbet bitti — yalnızca `includeExpired=true` ile döner.
        minutesUntilClose:
          type: integer
          description: Nöbetin bitmesine kalan dakika. Bitmişse negatif.
    PharmacyList:
      type: object
      required: [stale, dataAsOf, nextRotationAt, cityCode, items]
      additionalProperties: false
      properties:
        stale:
          type: boolean
          description: |
            `true` ise veri bugünün rotasyonuna ait değil. Arayüz çevrimdışı/kayıtlı
            bilgi bandını gösterir ve birincil aksiyonu "Ara" yapar.
        dataAsOf:
          type: [string, "null"]
          format: date-time
          description: Veriyi kaynaktan çeken son başarılı senkronun zamanı.
        nextRotationAt:
          type: string
          format: date-time
          description: Nöbetin döneceği bir sonraki an (her gün 08:00 TRT).
        cityCode: { type: integer, example: 7 }
        items:
          type: array
          items: { $ref: "#/components/schemas/Pharmacy" }
...